Split `PreviewPrompt` into its own file
