+++ client/web/src/enterprise/batches/create/workspaces-preview/WorkspacesPreview.module.scss
@@ -4,27 +4,3 @@
     padding-bottom: 0.5rem;
     border-bottom: 1px solid var(--border-color);
 }
-
-.preview-prompt {
-    &-icon {
-        height: 50px;
-        width: 50px;
-        border-radius: 50%;
-        border: 1px solid var(--border-color);
-        background-color: var(--white);
-    }
-
-    &-header {
-        margin-top: 1rem;
-        margin-bottom: 1rem;
-        font-weight: 300;
-        font-style: italic;
-        color: var(--text-muted);
-        text-align: center;
-    }
-
-    &-on-example {
-        width: 100%;
-        max-width: 380px;
-    }
-}
